requirements.txt
# PaddleOCR and Core Dependencies
paddlepaddle>=3.0.0
paddleocr>=3.0.0
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.21.0

# FastAPI for API server
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Additional OCR utilities
pytesseract>=0.3.10
easyocr>=1.7.0

# Image processing
scikit-image>=0.19.0
imgaug>=0.4.0

# MCP Server dependencies
mcp>=1.0.0
pydantic>=2.0.0
asyncio-mqtt>=0.13.0

# Utility libraries
requests>=2.31.0
aiofiles>=23.0.0
python-dotenv>=1.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0